User model is wrapper around Spotify User
We use the RSpotify gem's User class to hold the actual user details
from Spotify. This data is stored in the DB with the User model, as a
hash in JSON format, and updated whenever new data is received by the
OAuth authentication process.

The User model then is just a simple wrapper around the RSpotify::User
class.